refactor: remove domains directory and flatten structure
Move all directories from src/domains/ to src/ level for simpler organization:
- Move src/domains/analytics → src/analytics
- Move src/domains/layout → src/layout
- Move src/domains/search → src/search
- Remove empty domains directory

Update all import paths throughout codebase:
- Fix relative imports in moved directories (32 files updated)
- Update absolute imports in shared components
- Maintain all functionality while simplifying structure

Benefits:
- Shorter, more intuitive import paths
- UX-focused organization over domain-driven design
- Easier navigation and component discovery
- Consistent with frontend best practices

Build and tests pass successfully with new structure.
