+++ src/layout/Recommendations/RecommendationCard/Skills/Tags.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { SkillTags } from './Tags';
-import type { Recommendation } from '../../../../../data/recommendations';
-import type { SkillCategory } from '../../../../../data/skills';
+import type { Recommendation } from '../../../../data/recommendations';
+import type { SkillCategory } from '../../../../data/skills';
 
 // Mock recommendation with diverse skills across actual categories from the system
 const mockRecommendationWithCategorizedSkills: Recommendation = {
