+++ src/layout/Recommendations/RecommendationCard/Skills/Tags.tsx
@@ -2,8 +2,8 @@ import React, { useMemo } from 'react';
 
 
 import styles from '../component.module.css';
-import { getSkillCategory, type SkillCategory } from '../../../../../data/skills';
-import type { Recommendation } from '../../../../../data/recommendations';
+import { getSkillCategory, type SkillCategory } from '../../../../data/skills';
+import type { Recommendation } from '../../../../data/recommendations';
 
 interface SkillTagsProps {
   recommendation: Recommendation;
