+++ src/layout/Recommendations/RecommendationCard/component.module.css
