+++ src/layout/TechnicalImplementation/component.tsx
