+++ src/layout/TechnicalImplementation/index.ts
