+++ src/search/components/HighlightedText/HighlightedText.test.tsx
