+++ src/search/components/HighlightedText/HighlightedText.tsx
