+++ src/search/components/HighlightedText/word-highlighting-demo.test.tsx
