+++ src/search/components/SearchBar/SearchBar.module.css
@@ -117,7 +117,7 @@
   color: var(--color-text);
 }
 
-.searchFieldOption input[type="checkbox"] {
+.searchFieldOption input[type='checkbox'] {
   margin: 0;
   cursor: pointer;
 }
