+++ src/search/utils/textHighlighting.ts
